Added IP address block.
This block returns the IP address of the global Bee object
as a string.
The function is implemented as a part of the senseBox-MCU library.
It maybe therefore be a good idea to first publish a new version
of the senseBox library before using this block in production.
My editor also removed trailing white spaces and inconsistent uses of
tabs as indentation. :)
I hope that's not a problem.